Moved Constants file
