+++ library/src/main/java/nl/surfnet/nonweb/sso/SSOServiceActivity.java
@@ -26,6 +26,7 @@ import android.support.annotation.Nullable;
 import android.util.Log;
 
 import nl.surfnet.nonweb.sso.data.Credential;
+import nl.surfnet.nonweb.sso.util.Constants;
 
 
 /**
